--- library/core/registry/src/main/java/org/quiltmc/qsl/registry/api/event/RegistryEntryPredicate.java
@@ -1,11 +1,0 @@
-package org.quiltmc.qsl.registry.api.event;
-
-@FunctionalInterface
-public interface RegistryEntryPredicate<V> {
-	boolean test(RegistryIterationContext<V> context);
-
-	default RegistryEntryPredicate<V> and(RegistryEntryPredicate<V> other) {
-		return context ->
-				this.test(context) && other.test(context);
-	}
-}
