--- library/core/registry/src/main/java/org/quiltmc/qsl/registry/api/event/RegistryIterationContext.java
@@ -1,14 +1,0 @@
-package org.quiltmc.qsl.registry.api.event;
-
-import net.minecraft.util.Identifier;
-import net.minecraft.util.registry.Registry;
-
-public interface RegistryIterationContext<V> {
-	Registry<V> registry();
-
-	V entry();
-
-	Identifier id();
-
-	int rawId();
-}
