+++ library/core/registry/src/main/java/org/quiltmc/qsl/registry/impl/event/RegistryEventStorage.java
@@ -4,9 +4,18 @@ import net.minecraft.util.registry.Registry;
 import org.quiltmc.qsl.base.api.event.ArrayEvent;
 import org.quiltmc.qsl.registry.api.event.RegistryEvents;
 
+/**
+ * A duck interface for getting registry events stored in a {@link Registry}.
+ */
 public interface RegistryEventStorage<V> {
+	/**
+	 * Gets the entry added event.
+	 */
 	ArrayEvent<RegistryEvents.EntryAdded<V>> quilt$getEntryAddedEvent();
 
+	/**
+	 * Casts a {@link Registry} to the duck interface.
+	 */
 	@SuppressWarnings("unchecked")
 	static <W> RegistryEventStorage<W> as(Registry<W> registry) {
 		return (RegistryEventStorage<W>) registry;
