+++ library/core/registry/src/main/java/org/quiltmc/qsl/registry/mixin/SimpleRegistryAccessor.java
@@ -6,6 +6,10 @@ import net.minecraft.util.registry.SimpleRegistry;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
+/**
+ * Used to access the namespaced ID to entry map in a {@link net.minecraft.util.registry.Registry} to increase iteration
+ * speed.
+ */
 @Mixin(SimpleRegistry.class)
 public interface SimpleRegistryAccessor<V> {
 	@Accessor("idToEntry")
