+++ library/core/registry/src/main/java/org/quiltmc/qsl/registry/mixin/SimpleRegistryMixin.java
@@ -6,7 +6,7 @@ import net.minecraft.util.registry.RegistryKey;
 import net.minecraft.util.registry.SimpleRegistry;
 import org.quiltmc.qsl.base.api.event.ArrayEvent;
 import org.quiltmc.qsl.registry.api.event.RegistryEvents;
-import org.quiltmc.qsl.registry.impl.event.MutableRegistryIterationContextImpl;
+import org.quiltmc.qsl.registry.impl.event.MutableRegistryEntryContextImpl;
 import org.quiltmc.qsl.registry.impl.event.RegistryEventStorage;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
@@ -14,9 +14,12 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
+/**
+ * Invokes events for {@link Registry} manipulation.
+ */
 @Mixin(SimpleRegistry.class)
 public abstract class SimpleRegistryMixin<V> extends Registry<V> implements RegistryEventStorage<V> {
-	@Unique private final MutableRegistryIterationContextImpl<V> quilt$iterationContext = new MutableRegistryIterationContextImpl<>(this);
+	@Unique private final MutableRegistryEntryContextImpl<V> quilt$entryContext = new MutableRegistryEntryContextImpl<>(this);
 
 	@Unique
 	private final ArrayEvent<RegistryEvents.EntryAdded<V>> quilt$entryAddedEvent = ArrayEvent.create(RegistryEvents.EntryAdded.class, callbacks -> context -> {
@@ -34,8 +37,8 @@ public abstract class SimpleRegistryMixin<V> extends Registry<V> implements Regi
 			at = @At("HEAD")
 	)
 	private void quilt$invokeEntryAddEvent(int rawId, RegistryKey<V> key, V entry, Lifecycle lifecycle, boolean checkDuplicateKeys, CallbackInfoReturnable<V> cir) {
-		quilt$iterationContext.set(key.getValue(), entry, rawId);
-		quilt$entryAddedEvent.invoker().onAdded(quilt$iterationContext);
+		quilt$entryContext.set(key.getValue(), entry, rawId);
+		quilt$entryAddedEvent.invoker().onAdded(quilt$entryContext);
 	}
 
 	@Override
